Renamed some modal files
